requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssbeautifier==1.14.7
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-ckeditor-5==0.2.7
django-cleanup==7.0.0
django-colorfield==0.8.0
django-extensions==3.2.1
django-filter==23.2
django-imagekit==4.1.0
django-js-asset==2.0.0
djangorestframework==3.14.0
EditorConfig==0.12.3
executing==1.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==6.6.0
ipython==8.13.2
jedi==0.18.2
jsbeautifier==1.14.7
json5==0.9.14
Markdown==3.4.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.0
urllib3==2.0.3
wcwidth==0.2.6
zipp==3.15.0
